Watchdog; volume fix 3; heartbeat
Watchdog timer added (500ms)

Final (?!) fix for volume problem (explitit set before playing every
sample). This is somewhat giving up, but it seems like the DFPlayer Mini
is resetting at random times or something like that and there isn't time
to full investigate, or the potential to fix the hardware, so explicity
setting the volume every time seems like a reasonale compromise.

Set the ProjectHeartbeat to show when in brownout mode.