Minimum value detection
We have noise on the film (numbers) which are adding noise to the frame
detection.  However, the noise is generally in the 300-800 range, so we
will only trigger a frame if the low end of the leading edge is below
some threshold (set with SHUTTER_SYNC_LOW_TRIGGER).  It works well!