Allow some slippage during long mark (1) read
The film seems to slip sometimes, resulting in marks being seen as being
too long.  We can reduce the impact of this by allowing very long marks
for the longer (1) values, si I increased the long mark max length to be
very long. Won't catch slippage in shorter marks, but there's no easy
fix for that in software.